+++ gestion_base_adresse/processing/data/load_layers.py
@@ -4,22 +4,16 @@ __email__ = "info@3liz.org"
 __revision__ = "$Format:%H$"
 
 from qgis.core import (
-    Qgis,
     QgsDataSourceUri,
     QgsProcessingContext,
     QgsProcessingOutputMultipleLayers,
     QgsProcessingOutputString,
-    QgsProcessingParameterString,
     QgsProviderRegistry,
     QgsVectorLayer,
     QgsExpressionContextUtils,
+    QgsProcessingParameterDatabaseSchema,
+    QgsProcessingParameterProviderConnection,
 )
-
-if Qgis.QGIS_VERSION_INT >= 31400:
-    from qgis.core import (
-        QgsProcessingParameterDatabaseSchema,
-        QgsProcessingParameterProviderConnection,
-    )
 
 from ...definitions.variables import GESTION_ADRESSE_POINT_ADRESSE, GESTION_ADRESSE_VOIE
 from ...qgis_plugin_tools.tools.algorithm_processing import BaseProcessingAlgorithm
@@ -54,57 +48,25 @@ class LoadLayersAlgorithm(BaseProcessingAlgorithm):
     def initAlgorithm(self, config):
         # INPUTS
         # Database connection parameters
-        label = tr("Connexion PostgreSQL vers la base de données")
-        tooltip = "Base de données de destination"
         default = QgsExpressionContextUtils.globalScope().variable('adresse_connection_name')
-        if Qgis.QGIS_VERSION_INT >= 31400:
-            param = QgsProcessingParameterProviderConnection(
-                self.CONNECTION_NAME,
-                label,
-                "postgres",
-                optional=False,
-                defaultValue=default
-            )
-        else:
-            param = QgsProcessingParameterString(self.CONNECTION_NAME, label, defaultValue=default)
-            param.setMetadata(
-                {
-                    "widget_wrapper": {
-                        "class": "processing.gui.wrappers_postgis.ConnectionWidgetWrapper"
-                    }
-                }
-            )
-        if Qgis.QGIS_VERSION_INT >= 31600:
-            param.setHelp(tooltip)
-        else:
-            param.tooltip_3liz = tooltip
+        param = QgsProcessingParameterProviderConnection(
+            self.CONNECTION_NAME,
+            tr("Connexion PostgreSQL vers la base de données"),
+            "postgres",
+            optional=False,
+            defaultValue=default
+        )
+        param.setHelp("Base de données de destination")
         self.addParameter(param)
 
-        label = tr("Schéma")
-        tooltip = 'Nom du schéma des données adresses'
-        default = 'adresse'
-        if Qgis.QGIS_VERSION_INT >= 31400:
-            param = QgsProcessingParameterDatabaseSchema(
-                self.SCHEMA,
-                label,
-                self.CONNECTION_NAME,
-                defaultValue=default,
-                optional=False,
-            )
-        else:
-            param = QgsProcessingParameterString(self.SCHEMA, label, default, False, True)
-            param.setMetadata(
-                {
-                    "widget_wrapper": {
-                        "class": "processing.gui.wrappers_postgis.SchemaWidgetWrapper",
-                        "connection_param": self.CONNECTION_NAME,
-                    }
-                }
-            )
-        if Qgis.QGIS_VERSION_INT >= 31600:
-            param.setHelp(tooltip)
-        else:
-            param.tooltip_3liz = tooltip
+        param = QgsProcessingParameterDatabaseSchema(
+            self.SCHEMA,
+            tr("Schéma"),
+            self.CONNECTION_NAME,
+            defaultValue='adresse',
+            optional=False,
+        )
+        param.setHelp('Nom du schéma des données adresses')
         self.addParameter(param)
 
         # OUTPUTS
@@ -144,23 +106,18 @@ class LoadLayersAlgorithm(BaseProcessingAlgorithm):
         layers_name_none["document"] = ""
         layers_name_none["v_commune"] = "insee_code"
 
-        if Qgis.QGIS_VERSION_INT >= 31400:
-            connection_name = self.parameterAsConnectionName(parameters, self.CONNECTION_NAME, context)
-            schema = self.parameterAsSchema(parameters, self.SCHEMA, context)
-        else:
-            connection_name = self.parameterAsString(parameters, self.CONNECTION_NAME, context)
-            schema = self.parameterAsString(parameters, self.SCHEMA, context)
+        connection_name = self.parameterAsConnectionName(parameters, self.CONNECTION_NAME, context)
+        schema = self.parameterAsSchema(parameters, self.SCHEMA, context)
 
         feedback.pushInfo("## CONNEXION A LA BASE DE DONNEES ##")
 
-        uri = None
         metadata = QgsProviderRegistry.instance().providerMetadata('postgres')
         connection = metadata.findConnection(connection_name)
         uri = QgsDataSourceUri(connection.uri())
 
         is_host = uri.host() != ""
         if is_host:
-            feedback.pushInfo("Connexion établie via l'hote")
+            feedback.pushInfo("Connexion établie via l'hôte")
         else:
             feedback.pushInfo("Connexion établie via le service")